Move RegisterFromFile to ItemTypeRegister.cs
